Front end register/ confirm before logout

+++ assets/script/home.js
@@ -19,6 +19,8 @@ function init() {
 // logout button delete cookie
 const logout = document.getElementById("logout");
 logout.addEventListener("click", () => {
-  cookie.deleteCookie("name");
-  window.location.reload();
+  if (window.confirm(`Do you want to logout?`)) {
+    cookie.deleteCookie("name");
+    window.location.reload();
+  }
 });
