+++ assets/script/upload.js
@@ -102,6 +102,8 @@ async function storeUploadedFileInCookie(file) {
 // logout button delete cookie
 const logout = document.getElementById("logout");
 logout.addEventListener("click", () => {
-  cookie.deleteCookie("name");
-  window.location.reload();
+  if (window.confirm(`Do you want to logout?`)) {
+    cookie.deleteCookie("name");
+    window.location.reload();
+  }
 });
